changed folders & styled logo
changed folders (have read better the assignment's requirements) and
added two pictures after modifying on photoshop. One will be used as a
logo on the main title and the other as the backround of the site. In
the css file, there is a rule added that styles the main title (logo).
One font from google fonts is added to serve that cause