Bio is styled; CV is created. Footers added.
Styled the bio page. CV page is created and content is added in the form
of lists. It doesn't yet looks like what i have in mind but soon it
will. No style for the CV page yet, working on that. Footers are also
added to every page, and also styled as well.